pyproject.toml
[project]
name = "pyserver-manager"
version = "0.1.0"
description = "Interactive CLI to orchestrate lab applications"
readme = "README.md"
authors = [{ name = "Your Name" }]
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra"
